refactor: extract generation studio composable
